refactor: removed base classes
